+++ src/input_streams.rs
@@ -34,9 +34,9 @@ pub struct InputStreams<'a> {
     /// A list of registered gamepads
     pub gamepads: &'a Gamepads,
     /// A [`KeyCode`] [`Input`] stream
-    pub keyboard: &'a Input<KeyCode>,
+    pub keycode: &'a Input<KeyCode>,
     /// A [`MouseButton`] [`Input`] stream
-    pub mouse: &'a Input<MouseButton>,
+    pub mouse_button: &'a Input<MouseButton>,
     /// A [`MouseWheel`] event stream
     pub mouse_wheel: &'a Events<MouseWheel>,
     /// A [`MouseMotion`] event stream
@@ -63,8 +63,8 @@ impl<'a> InputStreams<'a> {
             gamepad_button_axes,
             gamepad_axes,
             gamepads,
-            keyboard,
-            mouse,
+            keycode: keyboard,
+            mouse_button: mouse,
             mouse_wheel,
             mouse_motion,
             associated_gamepad: gamepad,
@@ -142,8 +142,8 @@ impl<'a> InputStreams<'a> {
                     false
                 }
             }
-            InputKind::Keyboard(keycode) => self.keyboard.pressed(keycode),
-            InputKind::Mouse(mouse_button) => self.mouse.pressed(mouse_button),
+            InputKind::Keyboard(keycode) => self.keycode.pressed(keycode),
+            InputKind::Mouse(mouse_button) => self.mouse_button.pressed(mouse_button),
             InputKind::MouseWheel(mouse_wheel_direction) => {
                 let mut total_mouse_wheel_movement = 0.0;
 
@@ -371,12 +371,12 @@ pub struct MutableInputStreams<'a> {
     pub gamepad_events: &'a mut Events<GamepadEvent>,
 
     /// A [`KeyCode`] [`Input`] stream
-    pub keyboard: &'a mut Input<KeyCode>,
+    pub keycode: &'a mut Input<KeyCode>,
     /// Events used for mocking keyboard-related inputs
     pub keyboard_events: &'a mut Events<KeyboardInput>,
 
     /// A [`MouseButton`] [`Input`] stream
-    pub mouse: &'a mut Input<MouseButton>,
+    pub mouse_button: &'a mut Input<MouseButton>,
     /// Events used for mocking [`MouseButton`] inputs
     pub mouse_button_events: &'a mut Events<MouseButtonInput>,
     /// A [`MouseWheel`] event stream
@@ -425,9 +425,9 @@ impl<'a> MutableInputStreams<'a> {
             gamepad_axes: gamepad_axes.into_inner(),
             gamepads: gamepads.into_inner(),
             gamepad_events: gamepad_events.into_inner(),
-            keyboard: keyboard.into_inner(),
+            keycode: keyboard.into_inner(),
             keyboard_events: keyboard_events.into_inner(),
-            mouse: mouse.into_inner(),
+            mouse_button: mouse.into_inner(),
             mouse_button_events: mouse_button_events.into_inner(),
             mouse_wheel: mouse_wheel.into_inner(),
             mouse_motion: mouse_motion.into_inner(),
@@ -456,8 +456,8 @@ impl<'a> From<MutableInputStreams<'a>> for InputStreams<'a> {
             gamepad_button_axes: &*(mutable_streams.gamepad_button_axes),
             gamepad_axes: &*(mutable_streams.gamepad_axes),
             gamepads: &*(mutable_streams.gamepads),
-            keyboard: &*(mutable_streams.keyboard),
-            mouse: &*(mutable_streams.mouse),
+            keycode: &*(mutable_streams.keycode),
+            mouse_button: &*(mutable_streams.mouse_button),
             mouse_wheel: &*(mutable_streams.mouse_wheel),
             mouse_motion: &*(mutable_streams.mouse_motion),
             associated_gamepad: mutable_streams.associated_gamepad,
@@ -474,8 +474,8 @@ impl<'a> From<&'a MutableInputStreams<'a>> for InputStreams<'a> {
             gamepad_button_axes: &*(mutable_streams.gamepad_button_axes),
             gamepad_axes: &*(mutable_streams.gamepad_axes),
             gamepads: &*(mutable_streams.gamepads),
-            keyboard: &*(mutable_streams.keyboard),
-            mouse: &*(mutable_streams.mouse),
+            keycode: &*(mutable_streams.keycode),
+            mouse_button: &*(mutable_streams.mouse_button),
             mouse_wheel: &*(mutable_streams.mouse_wheel),
             mouse_motion: &*(mutable_streams.mouse_motion),
             associated_gamepad: mutable_streams.associated_gamepad,